requirements.txt
Flask==2.2.5
python-dotenv
mysql-connector-python==8.0.33
